% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run app}
\usage{
run_app()
}
\value{
No return value. This function starts a Shiny application.
}
\description{
Launches the Shiny web application using the predefined UI and server functions.
}
