% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyPCA.R
\name{runShinyPCA}
\alias{runShinyPCA}
\title{R Shiny Application for Principal Component Analysis}
\usage{
runShinyPCA()
}
\value{
Shiny App
}
\description{
R Shiny Application for Principal Component Analysis
}
\examples{
runShinyPCA()
}
