% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{sidebarTags}
\alias{sidebarTags}
\title{Sidebar UI Elements for File Upload}
\usage{
sidebarTags(id)
}
\arguments{
\item{id}{A namespace ID for the sidebar elements}
}
\value{
A list of UI elements for file upload and settings
}
\description{
Sidebar UI Elements for File Upload
}
