% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{uploadHeadData}
\alias{uploadHeadData}
\title{Main UI Plot for Upload page}
\usage{
uploadHeadData(id)
}
\arguments{
\item{id}{A namespace ID for the main elements}
}
\value{
A list of UI elements for file upload and settings
}
\description{
Main UI Plot for Upload page
}
