% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{uploadServer}
\alias{uploadServer}
\title{Upload Server Module}
\usage{
uploadServer(id)
}
\arguments{
\item{id}{A namespace ID for the module server logic}
}
\value{
A reactive dataset for use in the server logic
}
\description{
Upload Server Module
}
