% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{uploadUI}
\alias{uploadUI}
\title{Upload UI Module UI}
\usage{
uploadUI(id)
}
\arguments{
\item{id}{A namespace ID for the module UI}
}
\value{
UI for the upload module
}
\description{
Upload UI Module UI
}
