% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_welcome.R
\name{welcomeUI}
\alias{welcomeUI}
\title{Welcome UI Module}
\usage{
welcomeUI(id)
}
\arguments{
\item{id}{A namespace ID for the module}
}
\value{
A UI component (card) with welcome text
}
\description{
Welcome UI Module
}
